+++ merged-packages/smart-transactions-controller/src/index.test.ts
@@ -1,21 +1,25 @@
 import { ControllerMessenger } from '@metamask/base-controller';
+import {
+  type NetworkControllerGetNetworkClientByIdAction,
+  type NetworkControllerGetStateAction,
+  type NetworkControllerStateChangeEvent,
+} from '@metamask/network-controller';
 
 import DefaultExport, {
   type SmartTransactionsControllerActions,
   type SmartTransactionsControllerEvents,
 } from '.';
-import SmartTransactionsController, {
-  type AllowedActions,
-  type AllowedEvents,
-} from './SmartTransactionsController';
+import SmartTransactionsController from './SmartTransactionsController';
 import { ClientId } from './types';
 
 describe('default export', () => {
   it('exports SmartTransactionsController', () => {
     jest.useFakeTimers();
     const controllerMessenger = new ControllerMessenger<
-      SmartTransactionsControllerActions | AllowedActions,
-      SmartTransactionsControllerEvents | AllowedEvents
+      | SmartTransactionsControllerActions
+      | NetworkControllerGetNetworkClientByIdAction
+      | NetworkControllerGetStateAction,
+      SmartTransactionsControllerEvents | NetworkControllerStateChangeEvent
     >();
     const messenger = controllerMessenger.getRestricted({
       name: 'SmartTransactionsController',